requirements.txt
fastapi==0.121.2
uvicorn==0.29.0
httpx==0.28.1
python-dotenv==1.2.1
openai==2.8.0
pydantic==2.12.4
pydantic_core==2.41.5
annotated-types==0.7.0
typing-inspection==0.4.2
typing_extensions==4.15.0
annotated-doc==0.0.4
httpcore==1.0.9
sniffio==1.3.1
python-multipart==0.0.20
tqdm==4.67.1
idna==3.11
h11==0.16.0
certifi==2025.11.12
starlette==0.49.3
distro==1.9.0
colorama==0.4.6
click==8.3.0
et_xmlfile==2.0.0
openpyxl==3.1.5
jiter==0.12.0
psycopg2-binary==2.9.9
twilio==9.3.1

